requirements.txt
panda
mlflow
notebook
scikit-learn
matplotlib
pyYAML
tqdm
ensure==1.0.2
python-box==6.0.2
joblib
types-pyYAML
Flask
Flask-Cors
requests
-e .